move login session state storing from globalsetup to a fixture to allow tests to be written from perspective of a patient user. This is following latest guidance from playwright docs
