+++ tests/e2e/playwright/fixtures.ts
@@ -1,0 +1,46 @@
+// playwright/fixtures.ts
+import fs from 'fs';
+import path from 'path';
+
+import { test as baseTest } from '@playwright/test';
+
+import { baseURL } from "../../../playwright.config";
+
+export * from '@playwright/test';
+
+/**
+ * Fixture for writing tests within the browser context of a logged in patient.
+ * Do not use this fixture if your test modifies server-side state!
+ */
+export const patientTest = baseTest.extend<object, { workerStorageState: string }>({
+  storageState: ({ workerStorageState }, use) => use(workerStorageState),
+  
+  workerStorageState: [async ({ browser }, use, workerInfo) => {
+    const id = workerInfo.workerIndex;
+    const fileName = path.resolve(workerInfo.project.outputDir, `.auth/${id}.json`);
+
+    if (fs.existsSync(fileName)) {
+      await use(fileName); // Reuse existing authentication state if any.
+      return;
+    }
+
+    // authenticate in a clean environment by unsetting storage state.
+    const page = await browser.newPage({ storageState: undefined });
+
+    // Login as patient.
+    await page.goto(`${baseURL}/login`);
+    await page.locator("input[name=username]").fill("e2e-patient");
+    await page.locator("input[name=password]").fill("password");
+    const pageLoaded = page.waitForEvent("load");
+    await page.getByRole("button", { name: "Login" }).click();
+    await pageLoaded;
+    await page.goto(`${baseURL}/`, {
+      waitUntil: "domcontentloaded",
+    });
+
+    // Save state
+    await page.context().storageState({ path: fileName });
+    await page.close();
+    await use(fileName);
+  }, { scope: 'worker' }],
+});
