+++ tests/e2e/setup/globalSetup.ts
@@ -1,9 +1,6 @@
-import type { BrowserType } from "@playwright/test";
-import { chromium, firefox, webkit } from "@playwright/test";
 import { Role } from "@prisma/client";
 import * as argon2 from "argon2";
 
-import { baseURL } from "../../../playwright.config";
 import { prisma } from "../../../src/server/db";
 
 /**
@@ -57,39 +54,6 @@ async function globalSetup() {
     },
     update: {},
   });
-
-  //const browsers = [await chromium.launch(), await webkit.launch(), await firefox.launch()];
-  const browsers: Record<string, BrowserType> = {
-    chromium: chromium,
-    webkit: webkit,
-    firefox: firefox,
-  };
-
-  for (const name in browsers) {
-    const browserType: BrowserType | undefined = browsers[name];
-
-    if (!browserType) {
-      continue;
-    }
-
-    const browser = await browserType.launch();
-    const page = await browser.newPage();
-
-    await page.goto(`${baseURL}/login`);
-    await page.locator("input[name=username]").fill("e2e");
-    await page.locator("input[name=password]").fill("password");
-    const pageLoaded = page.waitForEvent("load");
-    await page.getByRole("button", { name: "Login" }).click();
-    await pageLoaded;
-    await page.goto(`${baseURL}/`, {
-      waitUntil: "domcontentloaded",
-    });
-
-    await page
-      .context()
-      .storageState({ path: `playwright/.auth/${name}/user.json` });
-    await browser.close();
-  }
 }
 
 export default globalSetup;
